+++ packages/eas-cli/src/credentials/ios/actions/BuildCredentialsUtils.ts
@@ -91,9 +91,9 @@ export async function assignBuildCredentialsAsync(
 }
 
 export async function getAppFromContextAsync(ctx: CredentialsContext): Promise<App> {
-  ctx.ensureProjectContext();
-  const projectName = ctx.exp.slug;
-  const projectId = ctx.projectId;
+  const exp = await ctx.getExpoConfigAsync();
+  const projectName = exp.slug;
+  const projectId = await ctx.getProjectIdAsync();
   const account = await getOwnerAccountForProjectIdAsync(ctx.graphqlClient, projectId);
   return {
     account,