+++ packages/eas-cli/src/project/ios/entitlements.ts
@@ -1,7 +1,9 @@
-import { IOSConfig, compileModsAsync } from '@expo/config-plugins';
+import { ExportedConfig, IOSConfig, compileModsAsync } from '@expo/config-plugins';
 import { JSONObject } from '@expo/json-file';
 import { getPrebuildConfigAsync } from '@expo/prebuild-config';
+import spawnAsync from '@expo/spawn-async';
 
+import Log from '../../log';
 import { readPlistAsync } from '../../utils/plist';
 import { Client } from '../../vcs/vcs';
 import { hasIgnoredIosProjectAsync } from '../workflow';
@@ -10,6 +12,9 @@ interface Target {
   buildConfiguration?: string;
   targetName: string;
 }
+
+let wasExpoConfigPluginsWarnPrinted = false;
+
 export async function getManagedApplicationTargetEntitlementsAsync(
   projectDir: string,
   env: Record<string, string>,
@@ -22,14 +27,39 @@ export async function getManagedApplicationTargetEntitlementsAsync(
       ...process.env,
       ...env,
     };
-    const { exp } = await getPrebuildConfigAsync(projectDir, { platforms: ['ios'] });
 
-    const expWithMods = await compileModsAsync(exp, {
-      projectRoot: projectDir,
-      platforms: ['ios'],
-      introspect: true,
-      ignoreExistingNativeFiles: await hasIgnoredIosProjectAsync(projectDir, vcsClient),
-    });
+    let expWithMods: ExportedConfig;
+    try {
+      const { stdout } = await spawnAsync(
+        'npx',
+        ['expo', 'config', '--json', '--type', 'introspect'],
+
+        {
+          cwd: projectDir,
+          env: {
+            ...process.env,
+            ...env,
+            EXPO_NO_DOTENV: '1',
+          },
+        }
+      );
+      expWithMods = JSON.parse(stdout);
+    } catch (err: any) {
+      if (!wasExpoConfigPluginsWarnPrinted) {
+        Log.warn(
+          `Failed to read the app config from the project using "npx expo config" command: ${err.message}.`
+        );
+        Log.warn('Falling back to the version of "@expo/config" shipped with the EAS CLI.');
+        wasExpoConfigPluginsWarnPrinted = true;
+      }
+      const { exp } = await getPrebuildConfigAsync(projectDir, { platforms: ['ios'] });
+      expWithMods = await compileModsAsync(exp, {
+        projectRoot: projectDir,
+        platforms: ['ios'],
+        introspect: true,
+        ignoreExistingNativeFiles: await hasIgnoredIosProjectAsync(projectDir, vcsClient),
+      });
+    }
     return expWithMods.ios?.entitlements ?? {};
   } finally {
     process.env = originalProcessEnv;